Refactoring of Converters into a Single File A new UrlConverter has been added to the resource dictionary in App.xaml. The file StringNullOrEmptyToVisibilityConverter.cs containing the implementation of the StringNullOrEmptyToVisibilityConverter has been removed. The file UrlConverter.cs containing the implementation of the UrlConverter has been removed. The declaration of the UrlConverter in the resources of UpdateBookPage.xaml has been removed. A new file Converters.cs has been created, which now contains the implementations of the UrlConverter and StringNullOrEmptyToVisibilityConverter converters.
