--- Utils/StringNullOrEmptyToVisibilityConverter.cs
@@ -1,21 +1,0 @@
-using System.Globalization;
-
-namespace MyApp.Utils
-{
-    public class StringNullOrEmptyToVisibilityConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is string str)
-            {
-                return !string.IsNullOrEmpty(str);
-            }
-            return false;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
