+++ Views/Books/UpdateBookPage.xaml
@@ -4,14 +4,8 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:viewmodel="clr-namespace:MyApp.ViewModels"
-    xmlns:converters="clr-namespace:MyApp.Utils"
     Title="Update Book"
     x:DataType="viewmodel:UpdateBookViewModel">
-    <ContentPage.Resources>
-        <ResourceDictionary>
-            <converters:UrlConverter x:Key="UrlConverter" />
-        </ResourceDictionary>
-    </ContentPage.Resources>
     <VerticalStackLayout Margin="20">
         <Entry Placeholder="Title" Text="{Binding Book.Title}" />
         <Entry Placeholder="Author" Text="{Binding Book.Author}" />
